tests: Extracted createToken function to utils folder

+++ test/app.test.js
@@ -1,28 +1,40 @@
 const request = require('supertest');
 const app = require('server/app')
-const User = require(__root + 'components/user/model');
 const mongoose = require('mongoose');
+const User = require(__root + 'components/user/model');
+const createToken = require(__root + 'test/utils/createToken');
 
-beforeEach(() => {
+function dropDB(db = mongoose) {
+	if (!mongoose.connection) {
+		throw Error('Missing database connection');
+	}
 	return mongoose.connection.dropDatabase();
-});
+}
 
 describe('Root Path', () => {
 	test('It should be an Unauthorized request', async () => {
-		const response = await request(app).get('/');
+		const response = await request(app)
+			.get('/');
+
 		expect(response.statusCode).toBe(401);
+		// await dropDB();
 	});
 })
 
-
 describe('Return User', () => {
-	test('It should return Hello World!', async () => {
+	test('It should return user', async () => {
 		const user = new User({
 				givenName: 'narp',
 				email: 'narp@gmail.com',
 		});
 		user.save();
-		const response = await request(app).get(`/user/${user.id}`);
+
+		const token = createToken(user);
+		const response = await request(app)
+			.get(`/user/${user.id}`)
+			.set('Authorization', 'Bearer ' + token);
+		
+		expect(JSON.parse(response.text)["_id"]).toBe(user.id);
 	});
 })
